feat(ListPage): update the NanopubResults view mode handling to use NanodashSession
